pyproject.toml
[project]
name = "mlops-lab1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "mlflow>=3.7.0",
    "numpy>=2.3.4",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.6",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "pytorch-lightning>=2.6.0",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pylint>=4.0.3",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]
